add cuda fft fuzzer in honggfuzz
