--- fuzz/cuda_fuzz/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "cuda_fuzz"
-version.workspace = true
-edition.workspace = true
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-lambdaworks-math = { workspace = true, features = ["cuda"] }
-lambdaworks-gpu = { workspace = true, features = ["cuda"] }
-libfuzzer-sys = { workspace = true }
- 
-[[bin]]
-name = "cuda_fft_fuzzer"
-path = "fuzz_targets/cuda_fft_fuzzer.rs"
-test = false
-doc = false
