+++ fuzzer/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "fuzzer"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+honggfuzz = "0.5.51"
+lambdaworks-math = { path = "../math", features = ["cuda"] }
+lambdaworks-gpu = { path = "../gpu", features = ["cuda"] }
+
+
+[workspace]
+members = ["."]
+
+[[bin]]
+name = "cuda_fft_fuzzer"
+path = "src/cuda_fft_fuzzer.rs"
